requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
django-jsonfield==1.4.1
numpy==1.19.3
pandas==1.1.3
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
